this text file is for Ty
this is the homepage with a few corrections. All I did was position the home button and center a few things more precisely using positon: aboslute. There was also some repeat text that I got rid of. other than that good work. Also, good work on the formatting on the code, it was very easy to read and edit and that's most important so keep that up.